requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.6
django-toolbelt==0.0.1
django-allauth==0.43.0
django-bootstrap4==2.3.1
dj3-cloudinary-storage==0.0.3
gunicorn==20.0.4
psycopg2==2.8.6
pytz==2021.1
Pillow==7.0.0
sqlparse==0.4.1
static3==0.7.0
whitenoise==5.2.0
